+++ CHANGELOG.md
@@ -2,7 +2,6 @@
 
 ### Next Version
 * Partial support for timed metadata on Android MediaPlayer [#707](https://github.com/react-native-community/react-native-video/pull/707)
-* Update to ExoPlayer 2.8.2. Android SDK 26 now required [#1170](https://github.com/react-native-community/react-native-video/pull/1170)
 * Support video caching for iOS [#955](https://github.com/react-native-community/react-native-video/pull/955)
 * Video caching cleanups [#1172](https://github.com/react-native-community/react-native-video/pull/1172)
 * Add ipod-library support [#926](https://github.com/react-native-community/react-native-video/pull/926/files)
@@ -11,6 +10,8 @@
 * Remove unnecessary import in TextTrackType [#1229](https://github.com/react-native-community/react-native-video/pull/1229)
 * Prevent flash between poster and video [#1167](https://github.com/react-native-community/react-native-video/pull/1167)
 * Support react-native-dom [#1253](https://github.com/react-native-community/react-native-video/pull/1253)
+* Update to ExoPlayer 2.8.2. Android SDK 26 now required [#1170](https://github.com/react-native-community/react-native-video/pull/1170)
+* Update to ExoPlayer 2.8.4 [#1266](https://github.com/react-native-community/react-native-video/pull/1266)
 
 ### Version 3.2.0
 * Basic fullscreen support for Android MediaPlayer [#1138](https://github.com/react-native-community/react-native-video/pull/1138)